% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpar.R
\name{impliedNames}
\alias{impliedNames}
\title{Internal metrumrg function to generate names for outputs}
\usage{
impliedNames(x)
}
\arguments{
\item{x}{list of matrix for calculation}
}
\description{
Internal metrumrg function to generate names for outputs
}
