% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is.diagonal}
\alias{is.diagonal}
\title{Check if this is a diagonal matrix}
\usage{
is.diagonal(matrix, eps = 1e-12)
}
\arguments{
\item{matrix}{the object to check}

\item{eps}{tolerance for checking that off diagonals are zero}
}
\description{
Check if this is a diagonal matrix
}
