% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is.square.matrix}
\alias{is.square.matrix}
\title{Check if this is a square matrix}
\usage{
\method{is.square}{matrix}(x, ...)
}
\arguments{
\item{x}{matrix for testing}
}
\description{
Check if this is a square matrix
}
