% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{offdiag.halfmatrix}
\alias{offdiag.halfmatrix}
\title{Extract off diagonal elements of a a half matrix}
\usage{
\method{offdiag}{halfmatrix}(x, ...)
}
\arguments{
\item{x}{half matrix for calculation}
}
\description{
Extract off diagonal elements of a a half matrix
}
