% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ord.halfmatrix}
\alias{ord.halfmatrix}
\title{Calculate order of half matrix}
\usage{
\method{ord}{halfmatrix}(x, ...)
}
\arguments{
\item{x}{half matrix for calculation}
}
\description{
Calculate order of half matrix
}
