% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ord.matrix}
\alias{ord.matrix}
\title{Calculate order of a square matrix}
\usage{
\method{ord}{matrix}(x, ...)
}
\arguments{
\item{x}{matrix for testing}
}
\description{
Calculate order of a square matrix
}
