% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpar.R
\name{rinvchisq}
\alias{rinvchisq}
\title{Draw samples from a inverse chi-square distribution}
\usage{
rinvchisq(n, df, cov)
}
\arguments{
\item{n}{number of simulations to perform; passed to rchisq}

\item{df}{degrees of freedom; passed to rchisq}

\item{cov}{the matrix to simulate}
}
\description{
Draw samples from a inverse chi-square distribution
}
