% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpar.R
\name{riwish}
\alias{riwish}
\title{Draw sample from a inverse wishart distribution}
\usage{
riwish(s, df, prec)
}
\arguments{
\item{s}{pending}

\item{df}{degrees of freedom; passed to rgamma}

\item{prec}{pending}
}
\description{
Draw sample from a inverse wishart distribution
}
