% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpar.R
\name{riwish_diag}
\alias{riwish_diag}
\title{Implement riwish for a diagonal matrix
When we find that all off diagonal elements are 0, we will simulate
as a series of independent invchisq simulations}
\usage{
riwish_diag(n, df, cov)
}
\arguments{
\item{n}{number of simulations to perform; passed to rinvchisq}

\item{df}{degrees of freedom; passed to rinvchisq}

\item{cov}{the matrix to simulate}
}
\description{
Implement riwish for a diagonal matrix
When we find that all off diagonal elements are 0, we will simulate
as a series of independent invchisq simulations
}
