% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_matrix.R
\name{sblock}
\alias{sblock}
\title{metrumrg::simblock replacement that handles diagonal omega or sigma}
\usage{
sblock(n, df, cov)
}
\arguments{
\item{n}{number of simulations; scalar}

\item{df}{degrees of freedom; scalar}

\item{cov}{a single matrix to simulate}
}
\description{
metrumrg::simblock replacement that handles diagonal omega or sigma
}
