% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpar.R
\name{simblock}
\alias{simblock}
\title{Handle diagonal omega or sigma}
\usage{
simblock(n, df, cov, diagnal = FALSE)
}
\arguments{
\item{n}{number of simulations to perform}

\item{df}{degrees of freedom}

\item{cov}{the matrix to simulate}

\item{diagnal}{TRUE or FALSE; whether cov is a diagnal matrix or not}
}
\description{
Handle diagonal omega or sigma
}
