% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpar.R
\name{simpar}
\alias{simpar}
\title{Create Parameters for Simulation with Uncertainty}
\usage{
simpar(
  nsim,
  theta,
  covar,
  omega,
  sigma,
  odf = NULL,
  sdf = NULL,
  digits = 4,
  min = -Inf,
  max = Inf,
  omega_diag = FALSE,
  sigma_diag = FALSE,
  format = "df"
)
}
\arguments{
\item{nsim}{scalar numeric specifying the number of sets to attempt}

\item{theta}{vector of point estimates of fixed effect parameters}

\item{covar}{variance-covariance matrix for fixed effect parameters}

\item{omega}{list of variance-covariance matrices for first level random effects}

\item{sigma}{list of variance-covariance matrices for second level random effects}

\item{odf}{vector of omega degrees of freedom, one per matrix}

\item{sdf}{vector of sigma degrees of freedom, one per matrix}

\item{digits}{number of significant digits to include in output}

\item{min}{lower limit for parameter estimates}

\item{max}{upper limit for parameter estimates}

\item{omega_diag}{logical; default is FALSE, if TRUE, simulated omega off-diagonal elements will be 0}

\item{sigma_diag}{logical; default is FALSE, if TRUE, simulated sigma off-diagonal elements will be 0}

\item{format}{default is "df", which will provide output as a data frame, if "list", the output will be provided as a list}
}
\value{
matrix, with column names indicating parameters, and row names indicating
set number before filtering by min and max.
}
\description{
Create Parameters for Simulation with Uncertainty
}
\details{
If min or max are non-default (see below), you may want to set nsim
marginally higher to allow for dropped sets. covar is coerced to matrix
using as.matrix.

If omega and sigma are not lists, they are coerced using list. Then each
element is coerced using as.matrix.

By default, each element in odf and sdf will be the length (number of
elements) in the corresponding matrix.

\code{min} and \code{max} may be given as scalar values, in which case they apply to
all parameters (as do the defaults). Alternatively, the first n limits may
be specified as a vector, in which case the remaining (if any) will be the
default. If any simulated parameter does not fall between its limits, inclusive,
the entire parameter set (row) is dropped from the result, with warning.
}
\examples{

set.seed(100)
simpar(
 nsim=10,
 theta=c(13,75,1),
 covar=matrix(c(10,7,2,7,30,1,2,1,0.5),ncol=3,nrow=3),
 omega=list(
   0.1,
   matrix(c(0.04,0.02,0.02,0.04),ncol=2,nrow=2)
 ),
 odf=c(50,20),
 sigma=list(0.04,1),
 sdf=c(99,99),
 min=rep(0,3),
 max=rep(90,3)
)
simpar(
 nsim=1,
 theta=c(13,75,1),
 covar=matrix(c(10,7,2,7,30,1,2,1,0.5),ncol=3,nrow=3),
 omega=list(
   0.1,
   matrix(c(0.04,0.02,0.02,0.04),ncol=2,nrow=2)
 ),
 odf=c(50,20),
 sigma=list(0.04,1),
 sdf=c(99,99),
 min=rep(0,3),
 max=rep(90,3)
)
simpar(
 nsim=1,
 theta=c(13,75,1),
 covar=matrix(c(10,7,2,7,30,1,2,1,0.5),ncol=3,nrow=3),
 omega=list(
   0.1,
   matrix(c(0.04,0.02,0.02,0.04),ncol=2,nrow=2)
 ),
 odf=c(50,20),
 sigma=list(0.04,1),
 sdf=c(99,99),
 min=Inf,
 max=-1
)

}
