% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_matrix.R
\name{simulate_matrix}
\alias{simulate_matrix}
\title{Simulate OMEGA or SIGMA matrices}
\usage{
simulate_matrix(n, df, cov, prefix = "OM", style = c("simpar", "mrgsolve"))
}
\arguments{
\item{n}{number of simulations}

\item{df}{degrees of freedom; must be a list of numeric values with same
length as \code{cov}}

\item{cov}{a list of numeric matrices}

\item{prefix}{added to numbered outputs; only used when \code{style} is \code{"simpar"}}
}
\description{
Simulate OMEGA or SIGMA matrices
}
